% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{getPseudoBulk_malignant}
\alias{getPseudoBulk_malignant}
\title{Make pseudobulk data from spatial transcriptomics data.}
\usage{
getPseudoBulk_malignant(SpatialNeighborhood, STexpr)
}
\arguments{
\item{SpatialNeighborhood}{The spatial neighborhood data got with importFiles_SMTdb().}

\item{STexpr}{The spatial transcriptomcis expression data got with importFiles_SMTdb().}
}
\value{
A list of pseudobulk RNA counts, with each component represent a cancer type.
}
\description{
Aggregate the gene count across all Malignant/Non-malignant spots for each cancer sample.
}
\examples{
# getPseudoBulk(SpatialNeighborhood, STexpr)

}
