% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/script.R
\name{SampleSizeCircularBarplot}
\alias{SampleSizeCircularBarplot}
\title{Generate a circular barplot to summarize the sample size per cancer.}
\usage{
SampleSizeCircularBarplot(
  meta,
  title = "Pan-cancer tasks",
  OUT_DIR,
  h = 4,
  w = 4.5
)
}
\arguments{
\item{meta}{meta data table, it must at least have columns, 'Condition' and 'Cancer_type'.}

\item{title}{character. Plot title (default: "Pan-cancer tasks").}

\item{OUT_DIR}{character. Directory where the dotplot PDF will be written.}

\item{h}{numeric or NULL. Figure height in inches}

\item{w}{numeric or NULL. Figure width in inches}
}
\description{
Generate a circular barplot to summarize the sample size per cancer.
}
